Deleted unnecessary duplicate of python_interface.c
Now both Python 2 and 3 can work with a single python_interface.c file (all the Python3-specific functions work in Python2 as well, so the Python2-specific file was deleted).
Added library checks for glib-2.0, purple and python to the configure.ac file, so that the process is aborted if one of those is missing.