Structure changes
Now apertiumInterfaceAPY.py doesn't rely on apertiumFiles and can be used separately to interact with the APY.
apertiumFiles.py now allows the user to change the name of the file where the data will be stored. As a result of this, the 'pythonInit' function from python_interface.c now receives a parameter indicating the name of the file where the plugin preferences are stored.