New command and minor changes
Added a new command 'apertium_errors' that allows the user to enable and disable error notifications from the plugin at will.
Added more feedback notifications for commands to inform the user that the command was successful.
Some commands were not being unregistered on plugin unload. Now all the commands are properly unregistered.