Added new functions to python module
New functions in apertiumFIles.py to manipulate the dictionary without the need to retrieve it.
Updated Makefile so that installation of the python module is no longer part of the build. Instead, the module must now be installed before running the make command.
Update README to reflect this installation change.