Additional encoding fixes
Encoding problems (#1) resulting in malformed request URLs fixed.
Unified python code to parse the text to be translated so that it is not version dependent. Also, now only the message itself is parsed, instead of the full request. This way the use of the quote_plus 'safe' argument is avoided (this argument was used before to avoid parsing the '&' or '|' characters, which were part of the request, but it would also ignore the parsing of those characters inside the message to be translated).